added validation

+++ ideas-app/src/idea/idea.service.ts
@@ -1,5 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { timingSafeEqual } from 'crypto';
 import { Repository } from 'typeorm';
 import { IdeaDTO } from './idea.dto';
 import { IdeaEntity } from './idea.entity';
@@ -17,25 +18,40 @@ export class IdeaService {
 
     }
 
-    async create(data:IdeaDTO) {
+    async create(data: IdeaDTO) {
         const idea = await this.ideaRepository.create(data);
         await this.ideaRepository.save(idea);
         return idea;
     }
 
     async read(id: string) {
-        return await this.ideaRepository.findOne({ where: { id } });
+        const idea = await this.ideaRepository.findOne({ where: { id } });
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND)
+
+        }
+
+        return idea;
     }
 
-    async update(id: string, data:Partial<IdeaDTO>) {
+    async update(id: string, data: Partial<IdeaDTO>) {
+        let idea = await this.ideaRepository.findOne({ where: { id } });
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.update({ id }, data);
+        idea = await this.ideaRepository.findOne({ where: { id } });
         return await this.ideaRepository.findOne({ id })
     }
 
 
     async destroy(id: string) {
+        const idea = await this.ideaRepository.findOne({ where: { id } });
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.delete({ id });
-        return { deleted: true };
+        return idea;
     }
